+++ src/lib/database.ts
@@ -50,6 +50,9 @@ async function updateLine(sql: string) {
 }
 
 async function updateDb() {
+  updateLine(
+    "ALTER TABLE `citizens` ADD `is_dangerous` varchar(255) DEFAULT '0' AFTER `b_status`;",
+  );
   updateLine("ALTER TABLE `officers` ADD `suspended` varchar(255) DEFAULT '0' AFTER `status`;");
   updateLine("ALTER TABLE `penal_codes` ADD `fine_amount` varchar(255) DEFAULT NULL AFTER `des`;");
   updateLine("ALTER TABLE `penal_codes` ADD `jail_time` varchar(255) DEFAULT NULL AFTER `des`;");