+++ docs/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 **If you need help updating/installing or have found any bugs, please join [our Discord server](https://discord.gg/eGnrPqEH7U) or open a [GitHub issue here](https://github.com/Dev-CasperTheGhost/snaily-cadv3/issues/new/choose)**
 
+## 2.0.9
+
+- New: set a citizen as `Dangerous`
+- New: able to specify time for `Jail time` (use `<amount> hours`, `<amount> minutes`, `<amount> seconds`, `<amount> milliseconds`)
+- Minor: fixed a bug in `RegisterWeaponModal`
+- Minor: Updated dependencies
+- Minor: small code improvements
+- Minor: added missing translation keys
+
 ## 2.0.8-patch1
 
 - Fix: fixed bug where a citizen could not delete a registered vehicle